Remove POCSAG support, call user event function
POCSAG is no longer supported, it requires a bit different
approach and might be easily handled in as separate project.

Now calls user defined function when data recieved. For HDLC
wait until packet is completed. For ZVEI and DTMF is send
each byte in separate call. This should be avoided for ZVEI in
future, because there is grouping (moustly by 5 digits).